Reorganize Bazel and BUCK parser code #1678
    * Add Bazel and BUCK parser code back to build.py since the parsers are just subclasses of StarlarkManifestPackage
    * Add Bazel and BUCK test data to build directory and update tests
