+++ tests/packagedcode/test_build.py
@@ -36,10 +36,24 @@ from packages_test_utils import PackageTester
 
 
 class TestBuild(PackageTester):
-    test_data_dir = os.path.join(os.path.dirname(__file__), 'data')
+    test_data_dir = os.path.join(os.path.dirname(__file__), 'data/build')
+
+    def test_end2end_scan_can_detect_bazel(self):
+        test_file = self.get_test_loc('bazel/end2end')
+        expected_file = self.get_test_loc('bazel/end2end-expected.json')
+        result_file = self.get_temp_file('results.json')
+        run_scan_click(['--package', test_file, '--json-pp', result_file])
+        check_json_scan(expected_file, result_file, regen=False)
+
+    def test_end2end_scan_can_detect_buck(self):
+        test_file = self.get_test_loc('buck/end2end')
+        expected_file = self.get_test_loc('buck/end2end-expected.json')
+        result_file = self.get_temp_file('results.json')
+        run_scan_click(['--package', test_file, '--json-pp', result_file])
+        check_json_scan(expected_file, result_file, regen=False)
 
     def test_build_get_package_resources(self):
-        test_loc = self.get_test_loc('build/get_package_resources')
+        test_loc = self.get_test_loc('get_package_resources')
         codebase = Codebase(test_loc)
         root = codebase.root
         expected = [
@@ -49,3 +63,39 @@ class TestBuild(PackageTester):
         ]
         results = [r.path for r in build.BaseBuildManifestPackage.get_package_resources(root, codebase)]
         assert expected == results
+
+    def test_BuckPackage_recognize(self):
+        test_file = self.get_test_loc('buck/parse/BUCK')
+        result_packages = build.BuckPackage.recognize(test_file)
+        expected_packages = [
+            build.BuckPackage(name='app'),
+            build.BuckPackage(name='app2'),
+        ]
+        compare_package_results(expected_packages, result_packages)
+
+    def test_BuckPackage_recognize_with_license(self):
+        test_file = self.get_test_loc('buck/parse/license/BUCK')
+        result_packages = build.BuckPackage.recognize(test_file)
+        expected_packages = [
+            build.BuckPackage(
+                name='app',
+                declared_license=['LICENSE'],
+            )
+        ]
+        compare_package_results(expected_packages, result_packages)
+
+
+def compare_package_results(expected, result):
+    # We don't want to compare `root_path`, since the result will always
+    # have a different `root_path` than the expected result
+    result_packages = []
+    for result_package in result:
+        r = result_package.to_dict()
+        r.pop('root_path')
+        result_packages.append(r)
+    expected_packages = []
+    for expected_package in expected:
+        e = expected_package.to_dict()
+        e.pop('root_path')
+        expected_packages.append(e)
+    assert expected_packages == result_packages
